refactor: reorganized contact-us-page related files into the pages/contact directory
